Changed int to uint8 otherwise error for lack of Int struct

+++ fsw/src/canadarm_app_msg.h
@@ -55,7 +55,7 @@ typedef struct
 typedef struct
 {
    CFE_MSG_CommandHeader_t CmdHeader;
-   int pose_id;
+   uint8 pose_id;
 } CanadarmAppCmd_t;
 
 /*
